Add compression to SL

+++ application/src/main/java/common/param/sl/DepositEvent.java
@@ -38,7 +38,7 @@ public class DepositEvent extends TxnEvent {
      * @param accountTransfer
      * @param bookEntryTransfer
      */
-    public DepositEvent(int bid, int pid, String bid_array, String partition_index, int num_of_partition,
+    public DepositEvent(long bid, int pid, String bid_array, String partition_index, int num_of_partition,
                         String accountId,
                         String bookEntryId,
                         long accountTransfer,
@@ -88,7 +88,7 @@ public class DepositEvent extends TxnEvent {
     }
 
     public DepositEvent cloneEvent() {
-        DepositEvent event = new DepositEvent((int) bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, accountId, bookEntryId, accountTransfer, bookEntryTransfer);
+        DepositEvent event = new DepositEvent(bid, pid, Arrays.toString(bid_array), Arrays.toString(partition_indexs), number_of_partitions, accountId, bookEntryId, accountTransfer, bookEntryTransfer);
         event.setTimestamp(this.timestamp);
         return event;
     }
